TD POC
TLA OMG!

Added sink property to Unit, allowing units to eat units that collide
with them. When this happens they get a signal and can do whatever they
want to the unit they have eaten - they can even pass through unharmed
if desired. This allows waypoints ending in sinks to have an effect
when/if units reach the sink.